Refactor to reuse the blazor-root element from interop
https://github.com/smapiot/piral/pull/367#discussion_r599428349
